+++ src/webpack-config.js
@@ -2,6 +2,7 @@
 
 const webpack = require('webpack');
 const path = require('path');
+const processCwd = process.cwd();
 
 const config = {
   target: 'node',
@@ -24,7 +25,10 @@ const config = {
   ],
 
   resolveLoader: {
-    modulesDirectories: [path.join(__dirname, '../node_modules')],
+    modulesDirectories: [
+      path.join(processCwd, 'node_modules'),
+      path.join(__dirname, '../node_modules')
+    ],
   },
 
   module: {
@@ -34,7 +38,32 @@ const config = {
         exclude: /(node_modules|bower_components|vendor)/,
         loader: 'babel',
         query: {
-          presets: ['es2015', 'react', 'stage-0']
+          presets: [
+            require.resolve('babel-preset-es2015'),
+            require.resolve('babel-preset-react')
+          ],
+          plugins: [
+
+            // legacy decorator plugin (this must be the first plugin)
+            require.resolve('babel-plugin-transform-decorators-legacy'),
+
+            // stage-0
+            require.resolve('babel-plugin-transform-do-expressions'),
+            require.resolve('babel-plugin-transform-function-bind'),
+
+            // stage-1 without the decorator plugin
+            require.resolve('babel-plugin-transform-class-constructor-call'),
+            require.resolve('babel-plugin-transform-class-properties'),
+            require.resolve('babel-plugin-transform-export-extensions'),
+
+            // stage-2
+            require.resolve('babel-plugin-syntax-trailing-function-commas'),
+            require.resolve('babel-plugin-transform-object-rest-spread'),
+
+            // stage-3
+            require.resolve('babel-plugin-transform-async-to-generator'),
+            require.resolve('babel-plugin-transform-exponentiation-operator'),
+          ]
         }
       }
     ]