+++ src/webpack.js
@@ -16,15 +16,24 @@ module.exports = {
   register(gulp, taskName, parameters) {
     const cleanUpTaskName = taskName + '-clean';
     gulp.task(cleanUpTaskName, (done) => {
-      del([parameters.output.path], { force: true }).then(() => {
-        done();
-      });
+      Promise.all([].concat(parameters.config).map(webpackConfig => {
+        return del([webpackConfig.output.path], { force: true });
+      })).then(() => done()).catch(e => done(e));
     });
 
     gulp.task(taskName, [cleanUpTaskName], (done) => {
-      webpack(parameters).run((err) => {
+      const bundler = webpack(parameters.config);
+      const cb = (err/*, stats*/) => {
+        //console.log(stats.toString(parameters.config.stats || {})); //eslint-disable-line no-console
+
         done(err);
-      });
+      };
+
+      if (parameters.watch) {
+        bundler.watch(200, cb);
+      } else {
+        bundler.run(cb);
+      }
     });
   }
 };